requirements.txt
fastapi
uvicorn
pandas
firebase-admin
azure-storage-blob
requests
python-dotenv
azure-ai-inference